+++ CHANGELOG.md
@@ -7,8 +7,10 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Added
 * [`jsx-key`]: added `checkKeyMustBeforeSpread` option for new jsx transform ([#2835][] @morlay)
+* [`jsx-newline`]: add new rule ([#2693][] @jzabala)
 
 [#2835]: https://github.com/yannickcr/eslint-plugin-react/pull/2835
+[#2693]: https://github.com/yannickcr/eslint-plugin-react/pull/2693
 
 ## [7.21.5] - 2020.10.19
 
@@ -3216,3 +3218,4 @@ If you're still not using React 15 you can keep the old behavior by setting the 
 [`jsx-no-script-url`]: docs/rules/jsx-no-script-url.md
 [`no-adjacent-inline-elements`]: docs/rules/no-adjacent-inline-elements.md
 [`function-component-definition`]: docs/rules/function-component-definition.md
+[`jsx-newline`]: docs/rules/jsx-newline.md